+++ app/src/main/java/com/shift/pizzadeliveryapp/data/converters/toPizzaCatalog.kt
@@ -1,0 +1,10 @@
+package com.shift.pizzadeliveryapp.data.converters
+
+import com.shift.pizzadeliveryapp.data.remote.dto.GetCatalogDto
+import com.shift.pizzadeliveryapp.domain.models.PizzaCatalog
+
+fun GetCatalogDto.toPizzaCatalog() : PizzaCatalog {
+    return PizzaCatalog(
+        catalog = catalog.map { it.toPizza() }
+    )
+}
