+++ app/src/main/java/com/shift/pizzadeliveryapp/data/remote/PizzaApi.kt
@@ -2,11 +2,15 @@ package com.shift.pizzadeliveryapp.data.remote
 
 import com.shift.pizzadeliveryapp.data.remote.dto.GetCatalogDto
 import retrofit2.http.GET
+import retrofit2.http.POST
 
 interface PizzaApi {
 
     @GET("pizza/catalog")
     suspend fun getPizza() : GetCatalogDto
+
+    @POST("auth/otp")
+    suspend fun getOtpCode()
 
 /*    @POST("pizza/payment")
     suspend fun postPayment() //тут еще что-то будет
