+++ app/src/main/java/com/shift/pizzadeliveryapp/data/remote/dto/Pizza_DoughsDto.kt
@@ -2,7 +2,6 @@ package com.shift.pizzadeliveryapp.data.remote.dto
 
 
 import com.google.gson.annotations.SerializedName
-import com.shift.pizzadeliveryapp.domain.models.Pizza_Dough
 
 data class Pizza_DoughsDto(
     @SerializedName("name")
@@ -10,10 +9,3 @@ data class Pizza_DoughsDto(
     @SerializedName("price")
     val price: Int
 )
-
-fun Pizza_DoughsDto.toPizza_Dough() : Pizza_Dough{
-    return Pizza_Dough(
-        name = name,
-        price = price
-    )
-}
