+++ app/src/main/java/com/shift/pizzadeliveryapp/data/remote/dto/Pizza_IngredientsDto.kt
@@ -2,7 +2,6 @@ package com.shift.pizzadeliveryapp.data.remote.dto
 
 
 import com.google.gson.annotations.SerializedName
-import com.shift.pizzadeliveryapp.domain.models.Pizza_Ingredients
 
 data class Pizza_IngredientsDto(
     @SerializedName("cost")
@@ -12,11 +11,3 @@ data class Pizza_IngredientsDto(
     @SerializedName("name")
     val name: String
 )
-
-fun Pizza_IngredientsDto.toPizza_Ingredients() : Pizza_Ingredients{
-    return Pizza_Ingredients(
-        cost = cost,
-        img = img,
-        name = name
-    )
-}
