+++ app/src/main/java/com/shift/pizzadeliveryapp/data/remote/dto/Pizza_SizesDto.kt
@@ -3,7 +3,6 @@ package com.shift.pizzadeliveryapp.data.remote.dto
 
 
 import com.google.gson.annotations.SerializedName
-import com.shift.pizzadeliveryapp.domain.models.Pizza_Size
 
 data class Pizza_SizesDto(
     @SerializedName("name")
@@ -21,10 +20,3 @@ data class Pizza_SizesDto(
         "LARGE" -> R.string.pizza_large.toString()
         else -> "Другой"
     }*/
-
-fun Pizza_SizesDto.toPizza_Size() : Pizza_Size{
-    return Pizza_Size(
-        name = name,
-        price = price
-    )
-}
