+++ app/src/main/java/com/shift/pizzadeliveryapp/data/remote/dto/Pizza_ToppingsDto.kt
@@ -2,7 +2,6 @@ package com.shift.pizzadeliveryapp.data.remote.dto
 
 
 import com.google.gson.annotations.SerializedName
-import com.shift.pizzadeliveryapp.domain.models.Pizza_Toppings
 
 data class Pizza_ToppingsDto(
     @SerializedName("cost")
@@ -12,11 +11,3 @@ data class Pizza_ToppingsDto(
     @SerializedName("name")
     val name: String
 )
-
-fun Pizza_ToppingsDto.toPizza_Toppings() : Pizza_Toppings {
-    return Pizza_Toppings(
-        cost = cost,
-        img = img,
-        name = name
-    )
-}
